chore: rename 'methods' folder to 'functions'
