+++ src/functions/getResult/index.ts
